Make weird issues happen less
Firstly, we now periodically ping every non-TCP-only client on the UDP
socket in an attempt to keep the connection alive; SAR 1.12.2.1 and
above should return this ping on UDP (although we don't check the
response, it's just a keep-alive). Secondly, the client is now allowed
to miss a (TCP) ping as long as they respond to the next, which should
hopefully stop the random disconnects that I don't understand from
happening.